refactor : move GameMessage to view/message
