+++ src/main/java/racing/view/RacingGameInputView.java
@@ -3,7 +3,7 @@ package racing.view;
 import racing.domain.Car;
 import racing.domain.RacingCars;
 import racing.generator.CarListGenerator;
-import racing.message.GameMessage;
+import racing.view.message.GameMessage;
 
 import java.util.List;
 import java.util.Scanner;
