+++ src/main/java/racing/view/message/GameMessage.java
@@ -1,4 +1,4 @@
-package racing.message;
+package racing.view.message;
 
 public enum GameMessage {
     CAR_NAME_INPUT_MESSAGE("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분)."),
